refactor(hosted): isolate projection activation routes
