+++ crates/connect-hosted-provider/src/http.rs
@@ -46,6 +46,7 @@ mod accounts;
 mod authentication;
 mod authority_import_files;
 mod files;
+mod projections;
 
 use accounts::account_routes;
 use authentication::{bearer, request_origin, request_proof};
@@ -54,6 +55,7 @@ use authority_import_files::{
     prepare_authority_import_file_part,
 };
 use files::file_routes;
+use projections::projection_routes;
 
 const MAX_BODY_BYTES: usize = 3 * 1024 * 1024;
 // Record imports are paged, but a page can contain several large canonical
@@ -109,18 +111,6 @@ struct RotateTokenRequest {
 #[derive(Debug, Deserialize)]
 struct CompactRequest {
     through: u64,
-}
-
-#[derive(Debug, Deserialize)]
-struct ActivateCandidateBRequest {
-    expected_head: u64,
-    expected_resource_revision: String,
-    confirmation: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct AdvanceProjectionRequest {
-    generation_id: Uuid,
 }
 
 #[derive(Debug, Deserialize)]
@@ -203,6 +193,7 @@ pub fn app(state: AppState) -> Router {
         .max_age(std::time::Duration::from_secs(600));
     let internal = Router::new()
         .merge(account_routes())
+        .merge(projection_routes())
         .route("/internal/v1/protocol-usage", get(protocol_usage))
         .route(
             "/internal/v1/collections/{collection_id}",
@@ -211,18 +202,6 @@ pub fn app(state: AppState) -> Router {
         .route(
             "/internal/v1/collections/{collection_id}/usage",
             get(collection_usage),
-        )
-        .route(
-            "/internal/v1/collections/{collection_id}/projection",
-            get(projection_status),
-        )
-        .route(
-            "/internal/v1/collections/{collection_id}/projection/activate-candidate-b",
-            post(activate_candidate_b),
-        )
-        .route(
-            "/internal/v1/collections/{collection_id}/projection/advance",
-            post(advance_projection),
         )
         .route(
             "/internal/v1/collections/{collection_id}/replicas",
@@ -426,59 +405,6 @@ async fn collection_usage(
     state.authorize_internal(&headers)?;
     let usage = state.provider.collection_usage(collection_id).await?;
     Ok(Json(json!({ "usage": usage })))
-}
-
-async fn projection_status(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Path(collection_id): Path<Uuid>,
-) -> ApiResult<Json<Value>> {
-    state.authorize_internal(&headers)?;
-    let status = state.provider.projection_status(collection_id).await?;
-    Ok(Json(json!({ "projection": status })))
-}
-
-async fn activate_candidate_b(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Path(collection_id): Path<Uuid>,
-    Json(input): Json<ActivateCandidateBRequest>,
-) -> ApiResult<Json<Value>> {
-    state.authorize_internal(&headers)?;
-    let expected_confirmation = format!(
-        "activate-candidate-b:{collection_id}:{}:{}",
-        input.expected_head, input.expected_resource_revision
-    );
-    if input.confirmation != expected_confirmation {
-        return Err(ApiError::bad_request(
-            "projection_activation_confirmation_invalid",
-            "The Candidate B activation confirmation does not match the expected binding.",
-        ));
-    }
-    let status = state
-        .provider
-        .request_candidate_b_activation(
-            collection_id,
-            input.expected_head,
-            input.expected_resource_revision,
-        )
-        .await?;
-    Ok(Json(json!({ "projection": status })))
-}
-
-async fn advance_projection(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Path(collection_id): Path<Uuid>,
-    Json(input): Json<AdvanceProjectionRequest>,
-) -> ApiResult<Json<Value>> {
-    state.authorize_internal(&headers)?;
-    let batch = state
-        .provider
-        .advance_projection_generation(collection_id, input.generation_id)
-        .await?;
-    let status = state.provider.projection_status(collection_id).await?;
-    Ok(Json(json!({ "batch": batch, "projection": status })))
 }
 
 async fn protocol_usage(
